refactor(web-authn): removes unnecessary hook
