+++ src/modules/auth/hooks/webAuthn/index.ts
@@ -1,4 +1,3 @@
-export * from './useWebAuthn';
 export * from './useWebAuthnForm';
 export * from './useWebAuthnInput';
 export * from './useWebAuthnLastLogin';
